requirements.txt
streamlit
pandas
openai
python-dotenv
pyyaml
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
polars
duckdb
tqdm
pytest
